pyproject.toml
[project]
name = "Telegram-Ingestion-Pipeline"
version = "0.1.0"
description = "A robust pipeline to extract text, media, events, and metadata from Telegram group chats and store them in a vector database, enabling retrieval-augmented generation (RAG) for LLMs based on group chat data."
requires-python = ">=3.10"
dependencies = [
    "pillow",
    "pytesseract",
    "python-telegram-bot",
    "dotenv",
    "pathlib",
    "opencv-python",
    "easyocr",
    "pdfplumber",
    "python-docx",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]